[android] interpolation update
adding the capability for a vector with mixed commands to be properly interpolated between each other.  in order to properly interpolate, we need to add the capability to know where the last point was in the vector, which is not supported by Path in android, so I introduced a KFPath wrapper object that only reveals methods that we use, as well as support some basic last point behavior.  tested on reaction faces and later with newton/wave