Update AfterEffects and iOS Keyframes to Support XY Rotation After Effects Script: Added in functionality to support the XY rotation on a 3D layer in After Effects. If you perform the transformation on the  of the  you will be able to animate the 3-D axes also. Currently the support is only for Z-axis rotation, so this PR will add in the new functionality. In the Keyframes file, once exported, the keys used for X and Y rotation are  and , respectively.
iOS SDK: In order to interpret and handle the new 3D layer transformations, the iOS SDK will now be able to intepret the values and insert the animations accordingly. Using a  new  enumeration on the  class, you are able to specify which axis the rotation animation should occur on. This will hopefully help abstract some of the code in that file, to prevent duplication.